+++ addon/components/object-list-view.js
@@ -1120,7 +1120,21 @@ export default FlexberryBaseComponent.extend(
           objectlistviewEventsService.updateWidthTrigger(componentName);
         }
       });
-    }
+    },
+
+    /**
+      Checks if "Enter" button was pressed.
+      If "Enter" button was pressed then apply filters.
+
+      @method actions.keyDownFilterAction
+    */
+    keyDownFilterAction(e) {
+      if (e.keyCode === 13) {
+        this.get('currentController').send('refreshList', this.get('componentName'));
+        e.preventDefault();
+        return false;
+      }
+    },
   },
 
   /**
@@ -2599,7 +2613,9 @@ export default FlexberryBaseComponent.extend(
     let parent = Ember.$(settings.parent);
     let table = Ember.$(settings.table);
 
-    parent.append(table);
+    if (!parent.children(table).length) {
+      parent.append(table);
+    }
 
     Ember.$('.full.height .flexberry-content .ui.main.container').css('margin-bottom', '0');