+++ tests/dummy/app/routes/components-examples/flexberry-lookup/numeric-autocomplete.js
@@ -1,0 +1,33 @@
+import EditFormRoute from 'ember-flexberry/routes/edit-form';
+import EditFormRouteOperationsIndicationMixin from 'ember-flexberry/mixins/edit-form-route-operations-indication';
+
+export default EditFormRoute.extend(EditFormRouteOperationsIndicationMixin, {
+  /**
+    Name of model projection to be used as record's properties limitation.
+
+    @property modelProjection
+    @type String
+    @default 'SuggestionE'
+   */
+  modelProjection: 'SuggestionEWithKarma',
+
+  /**
+    Name of model to be used as form's record type.
+
+    @property modelName
+    @type String
+    @default 'ember-flexberry-dummy-suggestion'
+   */
+  modelName: 'ember-flexberry-dummy-suggestion',
+
+  /**
+    Returns model related to current route.
+
+    @method model
+   */
+  model(params) {
+    let store = this.get('store');
+    let base = store.createRecord('ember-flexberry-dummy-suggestion');
+    return base;
+  }
+});